feat(judger/data)!: migrate to s3
The data service was a custom-built object storage. This commit replaced
the judger part with s3 for greater flexibility and space for future
extensions. The web part was remain unchanged, therefore the zip upload
feature is currently broken.

BREAKING CHANGES: data needs to be migrated to a s3-compatible service
manually. (I use minio for testing.) Simply upload all the zip files to
a s3 bucket (timestamps are not needed), and change the judger config.